// src/Infrastructure/Objects/AppIoContainer/enums/IocRegisteredObject.ts
export enum IocRegisteredObject {
	LOGGER = 'LOGGER',
	MQTT_CLIENT = 'MQTT_CLIENT',
	MQTT_CLIENT_CONNECTION = 'MQTT_CLIENT_CONNECTION',
	APP_SETTINGS = 'APP_SETTINGS',
	HEALTH_PROVIDER_OBJECT_ACTIVITY_REGISTER = 'HEALTH_PROVIDER_OBJECT_ACTIVITY_REGISTER',
	HEALTH_PROVIDER_OBJECT_TO_VERIFY_AVAILABILITY = 'HEALTH_PROVIDER_OBJECT_TO_VERIFY_AVAILABILITY',
	BATTERY_HEALTH_PROVIDER = 'BATTERY_HEALTH_PROVIDER',
	APP_HEALTH_PROVIDER = 'APP_HEALTH_PROVIDER',
	REMOTE_COMMAND_PROVIDER = 'REMOTE_COMMAND_PROVIDER',
	MQTT_INCOME_MESSAGE_HANDLER = 'MQTT_INCOME_MESSAGE_HANDLER',
	POS_TERMINAL_LINK_WITH_BANK_ERROR_PROVIDER = 'POS_TERMINAL_LINK_WITH_BANK_ERROR_PROVIDER',
}